separate config and action code
