+++ modules/database/src/test/java/org/eclipse/xpanse/modules/database/serviceconfiguration/ServiceChangeDetailsEntityTest.java
@@ -14,8 +14,8 @@ import java.util.UUID;
 import org.eclipse.xpanse.modules.database.service.ServiceDeploymentEntity;
 import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsEntity;
 import org.eclipse.xpanse.modules.database.serviceorder.ServiceOrderEntity;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.AnsibleTaskResult;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.enums.ServiceChangeStatus;
+import org.eclipse.xpanse.modules.models.servicechange.AnsibleTaskResult;
+import org.eclipse.xpanse.modules.models.servicechange.enums.ServiceChangeStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
