+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/ServiceConfigurationManager.java
@@ -9,25 +9,18 @@ package org.eclipse.xpanse.modules.deployment;
 import jakarta.annotation.Resource;
 import jakarta.transaction.Transactional;
 import java.time.OffsetDateTime;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.xpanse.modules.database.resource.ServiceResourceEntity;
 import org.eclipse.xpanse.modules.database.service.ServiceDeploymentEntity;
 import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsEntity;
-import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsQueryModel;
 import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsStorage;
 import org.eclipse.xpanse.modules.database.serviceconfiguration.ServiceConfigurationEntity;
 import org.eclipse.xpanse.modules.database.serviceconfiguration.ServiceConfigurationStorage;
-import org.eclipse.xpanse.modules.database.serviceorder.ServiceOrderEntity;
-import org.eclipse.xpanse.modules.database.serviceorder.ServiceOrderStorage;
 import org.eclipse.xpanse.modules.database.servicetemplate.ServiceTemplateEntity;
 import org.eclipse.xpanse.modules.database.servicetemplate.ServiceTemplateStorage;
 import org.eclipse.xpanse.modules.database.utils.EntityTransUtils;
@@ -35,19 +28,12 @@ import org.eclipse.xpanse.modules.logging.CustomRequestIdGenerator;
 import org.eclipse.xpanse.modules.models.common.enums.UserOperation;
 import org.eclipse.xpanse.modules.models.service.deployment.DeployResource;
 import org.eclipse.xpanse.modules.models.service.enums.DeployResourceKind;
-import org.eclipse.xpanse.modules.models.service.enums.TaskStatus;
 import org.eclipse.xpanse.modules.models.service.order.ServiceOrder;
 import org.eclipse.xpanse.modules.models.service.order.enums.ServiceOrderType;
-import org.eclipse.xpanse.modules.models.service.order.exceptions.ServiceOrderNotFound;
 import org.eclipse.xpanse.modules.models.service.utils.ServiceConfigurationVariablesJsonSchemaGenerator;
 import org.eclipse.xpanse.modules.models.service.utils.ServiceConfigurationVariablesJsonSchemaValidator;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.AnsibleHostInfo;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.ServiceChangeRequest;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.ServiceConfigurationChangeResult;
 import org.eclipse.xpanse.modules.models.serviceconfiguration.ServiceConfigurationDetails;
 import org.eclipse.xpanse.modules.models.serviceconfiguration.ServiceConfigurationUpdate;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.enums.ServiceChangeStatus;
-import org.eclipse.xpanse.modules.models.serviceconfiguration.exceptions.ServiceChangeDetailsEntityNotFoundException;
 import org.eclipse.xpanse.modules.models.serviceconfiguration.exceptions.ServiceConfigurationInvalidException;
 import org.eclipse.xpanse.modules.models.serviceconfiguration.exceptions.ServiceConfigurationNotFoundException;
 import org.eclipse.xpanse.modules.models.servicetemplate.Ocl;
@@ -57,20 +43,13 @@ import org.eclipse.xpanse.modules.models.servicetemplate.ServiceChangeScript;
 import org.eclipse.xpanse.modules.models.servicetemplate.exceptions.ServiceTemplateNotRegistered;
 import org.eclipse.xpanse.modules.models.servicetemplate.utils.JsonObjectSchema;
 import org.eclipse.xpanse.modules.security.auth.UserServiceHelper;
-import org.hibernate.exception.LockTimeoutException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
 
 /** Bean to manage service configuration. */
 @Slf4j
 @Component
 public class ServiceConfigurationManager {
-
-    private static final String IP = "ip";
-    private static final String HOSTS = "hosts";
 
     @Resource private ServiceDeploymentEntityHandler serviceDeploymentEntityHandler;
 
@@ -81,8 +60,6 @@ public class ServiceConfigurationManager {
     @Resource private ServiceTemplateStorage serviceTemplateStorage;
 
     @Resource private DeployService deployService;
-
-    @Resource private ServiceOrderStorage serviceOrderStorage;
 
     @Resource private ServiceChangeDetailsManager serviceChangeDetailsManager;
 
@@ -142,7 +119,7 @@ public class ServiceConfigurationManager {
             }
             validate(serviceTemplateEntity, configurationUpdate);
             UUID orderId = CustomRequestIdGenerator.generateOrderId();
-            addServiceChangeDetails(
+            createServiceChangeRequestsInDatabase(
                     orderId,
                     serviceId,
                     serviceDeploymentEntity,
@@ -170,7 +147,7 @@ public class ServiceConfigurationManager {
         }
     }
 
-    private void addServiceChangeDetails(
+    private void createServiceChangeRequestsInDatabase(
             UUID orderId,
             UUID serviceId,
             ServiceDeploymentEntity serviceDeployment,
@@ -186,20 +163,14 @@ public class ServiceConfigurationManager {
         List<ServiceChangeScript> configManageScripts = configManage.getConfigManageScripts();
         List<ServiceChangeParameter> configurationParameters =
                 configManage.getConfigurationParameters();
-
-        List<ServiceChangeDetailsEntity> requests =
-                serviceChangeDetailsManager.getAllServiceChangeDetails(
-                        orderId,
-                        serviceDeployment,
-                        updateRequestMap,
-                        deployResourceMap,
-                        configManageScripts,
-                        configurationParameters,
-                        ServiceOrderType.CONFIG_CHANGE);
-
-        if (!CollectionUtils.isEmpty(requests)) {
-            serviceChangeDetailsStorage.saveAll(requests);
-        }
+        serviceChangeDetailsManager.createAndQueueAllServiceChangeRequests(
+                orderId,
+                serviceDeployment,
+                updateRequestMap,
+                deployResourceMap,
+                configManageScripts,
+                configurationParameters,
+                ServiceOrderType.CONFIG_CHANGE);
     }
 
     private void validate(
@@ -226,141 +197,8 @@ public class ServiceConfigurationManager {
                 jsonObjectSchema);
     }
 
-    /** Query pending service change request for agent. */
-    @Transactional
-    public ResponseEntity<ServiceChangeRequest> getPendingServiceChangeRequest(
-            UUID serviceId, String resourceName) {
-        try {
-            ServiceChangeDetailsEntity oldestRequest =
-                    getOldestServiceChangeDetails(serviceId, resourceName);
-            if (Objects.isNull(oldestRequest)) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
-            }
-            checkPermission(
-                    oldestRequest.getServiceDeploymentEntity(),
-                    UserOperation.VIEW_CONFIGURATIONS_OF_SERVICE);
-            List<DeployResource> deployResources =
-                    getDeployResources(serviceId, DeployResourceKind.VM);
-            if (Objects.isNull(oldestRequest.getResourceName())) {
-                validateConfigManager(
-                        serviceId, oldestRequest.getChangeHandler(), resourceName, deployResources);
-            }
-            updateServiceOrder(oldestRequest);
-            ServiceChangeDetailsEntity request =
-                    updateServiceChangeDetails(oldestRequest, resourceName);
-            if (Objects.isNull(request)) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
-            }
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body(getServiceConfigurationChangeRequest(request, deployResources));
-        } catch (LockTimeoutException e) {
-            log.error("Update service configuration update request object timed out", e);
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
-        }
-    }
-
-    private void updateServiceOrder(ServiceChangeDetailsEntity oldestRequest) {
-        ServiceOrderEntity serviceOrderEntity = oldestRequest.getServiceOrderEntity();
-        if (Objects.isNull(serviceOrderEntity)) {
-            String errorMsg =
-                    String.format(
-                            "Service order with service change details id %s not found.",
-                            oldestRequest.getId());
-            log.error(errorMsg);
-            throw new ServiceOrderNotFound(errorMsg);
-        }
-        serviceOrderEntity.setCompletedTime(OffsetDateTime.now());
-        serviceOrderEntity.setTaskStatus(TaskStatus.IN_PROGRESS);
-        serviceOrderStorage.storeAndFlush(serviceOrderEntity);
-    }
-
-    private ServiceChangeDetailsEntity updateServiceChangeDetails(
-            ServiceChangeDetailsEntity request, String resourceName) {
-        if (Objects.isNull(request.getResourceName())) {
-            request.setResourceName(resourceName);
-        }
-        request.setStatus(ServiceChangeStatus.PROCESSING);
-        return serviceChangeDetailsStorage.storeAndFlush(request);
-    }
-
-    private ServiceChangeDetailsEntity getOldestServiceChangeDetails(
-            UUID serviceId, String resourceName) {
-        ServiceChangeDetailsQueryModel model =
-                new ServiceChangeDetailsQueryModel(
-                        null, serviceId, null, null, ServiceChangeStatus.PENDING);
-        List<ServiceChangeDetailsEntity> requests =
-                serviceChangeDetailsStorage.listServiceChangeDetails(model);
-        if (CollectionUtils.isEmpty(requests)) {
-            return null;
-        }
-        Optional<ServiceChangeDetailsEntity> oldestRequestOptional =
-                requests.stream()
-                        .filter(
-                                request ->
-                                        Objects.isNull(request.getResourceName())
-                                                || resourceName.equals(request.getResourceName()))
-                        .filter(
-                                request ->
-                                        request.getServiceOrderEntity() != null
-                                                && request.getServiceOrderEntity().getStartedTime()
-                                                        != null)
-                        .min(
-                                Comparator.comparing(
-                                        request ->
-                                                request.getServiceOrderEntity().getStartedTime()));
-        return oldestRequestOptional.orElse(null);
-    }
-
-    private void validateConfigManager(
-            UUID serviceId,
-            String configManager,
-            String resourceName,
-            List<DeployResource> deployResources) {
-        Map<String, List<DeployResource>> resourceNameMap =
-                deployResources.stream()
-                        .collect(Collectors.groupingBy(DeployResource::getResourceName));
-        if (!resourceNameMap.containsKey(resourceName)) {
-            String errorMsg =
-                    String.format(
-                            "The service with serviceId %s does not contain "
-                                    + "a resource with resource_name %s",
-                            serviceId, resourceName);
-            log.error(errorMsg);
-            throw new ServiceConfigurationInvalidException(List.of(errorMsg));
-        }
-        deployResources.forEach(
-                deployResource -> {
-                    if (resourceName.equals(deployResource.getResourceName())) {
-                        if (!configManager.equals(deployResource.getGroupName())) {
-                            String errorMsg =
-                                    String.format(
-                                            "The service with serviceId %s does"
-                                                    + " not contain a group with resource_name %s "
-                                                    + "and  group_name %s",
-                                            serviceId, resourceName, configManager);
-                            log.error(errorMsg);
-                            throw new ServiceConfigurationInvalidException(List.of(errorMsg));
-                        }
-                    }
-                });
-    }
-
-    private ServiceChangeRequest getServiceConfigurationChangeRequest(
-            ServiceChangeDetailsEntity request, List<DeployResource> deployResources) {
-        ServiceChangeRequest serviceChangeRequest = new ServiceChangeRequest();
-        serviceChangeRequest.setChangeId(request.getId());
-        Optional<ServiceChangeScript> configManageScriptOptional = getConfigManageScript(request);
-        configManageScriptOptional.ifPresent(
-                serviceChangeScript ->
-                        serviceChangeRequest.setAnsibleScriptConfig(
-                                serviceChangeScript.getAnsibleScriptConfig()));
-        serviceChangeRequest.setConfigParameters(request.getProperties());
-        serviceChangeRequest.setAnsibleInventory(getAnsibleInventory(deployResources));
-        return serviceChangeRequest;
-    }
-
-    private Optional<ServiceChangeScript> getConfigManageScript(
-            ServiceChangeDetailsEntity request) {
+    /** Gets the config management script from service template. */
+    public Optional<ServiceChangeScript> getConfigManageScript(ServiceChangeDetailsEntity request) {
         ServiceTemplateEntity serviceTemplateEntity =
                 serviceTemplateStorage.getServiceTemplateById(
                         request.getServiceDeploymentEntity().getServiceTemplateId());
@@ -372,141 +210,24 @@ public class ServiceConfigurationManager {
             log.error(errMsg);
             throw new ServiceTemplateNotRegistered(errMsg);
         }
-        Optional<ServiceChangeScript> configManageScriptOptional =
-                serviceTemplateEntity
-                        .getOcl()
-                        .getServiceConfigurationManage()
-                        .getConfigManageScripts()
-                        .stream()
-                        .filter(
-                                serviceChangeScript -> {
-                                    String configManager = request.getChangeHandler();
-                                    return configManager != null
-                                            && configManager.equals(
-                                                    serviceChangeScript.getChangeHandler());
-                                })
-                        .findFirst();
-        return configManageScriptOptional;
-    }
-
-    private Map<String, Object> getAnsibleInventory(List<DeployResource> deployResources) {
-        Map<String, List<DeployResource>> deployResourceMap =
-                deployResources.stream()
-                        .collect(Collectors.groupingBy(DeployResource::getGroupName));
-        Map<String, Object> ansibleInventory = new HashMap<>();
-        deployResourceMap.forEach(
-                (groupName, deployResourceList) -> {
-                    Map<String, AnsibleHostInfo> hosts = new HashMap<>();
-                    deployResourceList.forEach(
-                            deployResource -> {
-                                AnsibleHostInfo ansibleHostInfo = new AnsibleHostInfo();
-                                ansibleHostInfo.setAnsibleHost(
-                                        deployResource.getProperties().get(IP));
-                                hosts.put(deployResource.getResourceName(), ansibleHostInfo);
-                            });
-                    Map<String, Map<String, AnsibleHostInfo>> resourceMap = new HashMap<>();
-                    resourceMap.put(HOSTS, hosts);
-                    ansibleInventory.put(groupName, resourceMap);
-                });
-        return ansibleInventory;
-    }
-
-    private List<DeployResource> getDeployResources(
-            UUID serviceId, DeployResourceKind resourceKind) {
-        ServiceDeploymentEntity deployedService =
-                serviceDeploymentEntityHandler.getServiceDeploymentEntity(serviceId);
-        Stream<ServiceResourceEntity> resourceEntities =
-                deployedService.getDeployResources().stream();
-        if (Objects.nonNull(resourceKind)) {
-            resourceEntities =
-                    resourceEntities.filter(
-                            resourceEntity ->
-                                    resourceEntity.getResourceKind().equals(resourceKind));
-        }
-        return EntityTransUtils.transToDeployResources(resourceEntities.toList());
+        return serviceTemplateEntity
+                .getOcl()
+                .getServiceConfigurationManage()
+                .getConfigManageScripts()
+                .stream()
+                .filter(
+                        serviceChangeScript -> {
+                            String configManager = request.getChangeHandler();
+                            return configManager != null
+                                    && configManager.equals(serviceChangeScript.getChangeHandler());
+                        })
+                .findFirst();
     }
 
     /**
-     * Method to update service change result.
-     *
-     * @param changeId id of the update request.
-     * @param result result of the service change request.
+     * updates configuration data after the configuration change request is processed successfully.
      */
-    public void updateServiceChangeResult(UUID changeId, ServiceConfigurationChangeResult result) {
-        ServiceChangeDetailsEntity request = serviceChangeDetailsStorage.findById(changeId);
-        if (Objects.isNull(request)
-                || !ServiceChangeStatus.PROCESSING.equals(request.getStatus())) {
-            String errorMsg =
-                    String.format(
-                            "Service change details with id %s , status %s not found",
-                            changeId, ServiceChangeStatus.PROCESSING);
-            log.error(errorMsg);
-            throw new ServiceChangeDetailsEntityNotFoundException(errorMsg);
-        }
-        if (result.getIsSuccessful()) {
-            request.setStatus(ServiceChangeStatus.SUCCESSFUL);
-        } else {
-            request.setStatus(ServiceChangeStatus.ERROR);
-            request.setResultMessage(result.getError());
-        }
-        request.setTasks(result.getTasks());
-        serviceChangeDetailsStorage.storeAndFlush(request);
-        updateServiceChangeResult(request);
-    }
-
-    private void updateServiceChangeResult(ServiceChangeDetailsEntity request) {
-        ServiceChangeDetailsQueryModel model =
-                new ServiceChangeDetailsQueryModel(
-                        request.getServiceOrderEntity().getOrderId(),
-                        request.getServiceDeploymentEntity().getId(),
-                        null,
-                        null,
-                        null);
-        List<ServiceChangeDetailsEntity> requests =
-                serviceChangeDetailsStorage.listServiceChangeDetails(model);
-
-        if (CollectionUtils.isEmpty(requests)) {
-            String errorMsg =
-                    String.format(
-                            "Service configuration change details with service "
-                                    + "id %s not found, ",
-                            request.getServiceDeploymentEntity().getId());
-            log.error(errorMsg);
-            throw new ServiceChangeDetailsEntityNotFoundException(errorMsg);
-        }
-        boolean isNeedUpdateServiceConfigurationChange =
-                requests.stream()
-                        .allMatch(
-                                changeDetails ->
-                                        changeDetails.getStatus() == ServiceChangeStatus.SUCCESSFUL
-                                                || changeDetails.getStatus()
-                                                        == ServiceChangeStatus.ERROR);
-        if (isNeedUpdateServiceConfigurationChange) {
-            ServiceOrderEntity entity = request.getServiceOrderEntity();
-            if (Objects.nonNull(entity)) {
-                boolean isAllSuccessful =
-                        requests.stream()
-                                .allMatch(
-                                        changeDetails ->
-                                                request.getStatus()
-                                                        == ServiceChangeStatus.SUCCESSFUL);
-                if (isAllSuccessful) {
-                    updateServiceOrderByResult(entity, TaskStatus.SUCCESSFUL);
-                    updateServiceConfiguration(request);
-                } else {
-                    updateServiceOrderByResult(entity, TaskStatus.FAILED);
-                }
-            }
-        }
-    }
-
-    private void updateServiceOrderByResult(ServiceOrderEntity entity, TaskStatus status) {
-        entity.setTaskStatus(status);
-        entity.setCompletedTime(OffsetDateTime.now());
-        serviceOrderStorage.storeAndFlush(entity);
-    }
-
-    private void updateServiceConfiguration(ServiceChangeDetailsEntity request) {
+    public void updateServiceConfiguration(ServiceChangeDetailsEntity request) {
         ServiceConfigurationEntity serviceConfigurationEntity =
                 request.getServiceDeploymentEntity().getServiceConfiguration();
         Map<String, Object> config = request.getServiceOrderEntity().getRequestBody();
