+++ modules/models/src/main/java/org/eclipse/xpanse/modules/models/servicechange/ServiceChangeRequest.java
@@ -3,7 +3,7 @@
  * SPDX-FileCopyrightText: Huawei Inc.
  */
 
-package org.eclipse.xpanse.modules.models.serviceconfiguration;
+package org.eclipse.xpanse.modules.models.servicechange;
 
 import java.util.Map;
 import java.util.UUID;
@@ -15,7 +15,7 @@ import org.eclipse.xpanse.modules.models.servicetemplate.AnsibleScriptConfig;
 public class ServiceChangeRequest {
 
     private UUID changeId;
-    private Map<String, Object> configParameters;
+    private Map<String, Object> serviceChangeParameters;
     private AnsibleScriptConfig ansibleScriptConfig;
     private Map<String, Object> ansibleInventory;
 }
